HZAM

Contents

DataAnalysis

Types

Functions

HZAM

Contents

Functions

Index

Mating

Functions

PlotData

Functions

Population

Types

Functions

Functions

HZAM.run_one_HZAM_sim — Method

run_one_HZAM_sim(
    w_hyb::Real, 
    S_AM::Real, 
    ecolDiff::Real, 
    intrinsic_R::Real; 
    <keyword arguments>
)

Run a single HZAM simulation.

Arguments

w_hyb::Real: the hybrid fitness.

S_AM::Real: the strength of assortative mating.

ecolDiff::Real: the ecological difference between the two species.

intrinsic_R::Real: the intrinsic growth rate.

K_total::Integer=20000: the carrying capacity of the environment.

max_generations::Integer=1000: the number of generations that the simulation will run for.

total_loci::Integer=6: the total number of loci in the genome.

female_mating_trait_loci=1:3: the loci specifying the female's mate preference.

male_mating_trait_loci=1:3: the loci specifying the male's mating trait.

competition_trait_loci=1:3: the loci specifying the ecological trait (used in fitness related to resource use).

hybrid_survival_loci=1:3: the loci specifying the probability of survival to adulthood.

survival_fitness_method:String="epistasis": the method used to calculate the probability of survival to adulthood.

per_reject_cost=0: the fitness loss of female per male rejected (due to search time, etc.). Can take values of 0 to 1.

sigma_disp=0.05: the standard deviation of the normal distribution determining how far offspring will disperse from their mothers.

sigma_comp=0.01: the standard deviation for the normal curve used in calculating local density.

do_plot=true: the program will display a plot of the locations and hybrid indices of every individual while the simulation is running if this is true.

plot_int=10: the interval (measured in generations) between updating the plot.

gene_plot=false: if true, generates phenotype plots.

save_plot=false: if true, saves each plot to a PNG file.

track_spatial_data=false: if true, keeps track of bimodality, gene flow, variance, cline widths, and cline positions.

track_population_data=true: if true, keeps track of population size, hybridness, overlap, and hybrid zone width.

track_fitness=false: if true, keeps track of the average number of offspring for each phenotype.

track_mating_success=false: if true, keeps track of the average number of mates per male for each phenotype.

track_phenotypes=false: if true, keeps track of the number of individuals with each phenotype.

HZAM.run_HZAM_set — Method

run_one_HZAM_sim(
    w_hyb::Real, 
    S_AM::Real, 
    ecolDiff::Real, 
    intrinsic_R::Real; 
    <keyword arguments>
)

Run the simulation for 64 combinations of hybrid fitness and assortative mating strength and store the outcome of each simulation in a JLD2 file.

Arguments

set_name::String: the name assigned to the set of simulations.

intrinsic_R::Real: the intrinsic growth rate.

ecolDiff::Real: the ecological difference between the two species. Between 0 and 1.

K_total::Integer=20000: the carrying capacity of the environment.

max_generations::Integer=1000: the number of generations that the simulation will run for.

total_loci::Integer=6: the total number of loci in the genome.

female_mating_trait_loci=1:3: the loci specifying the female's mate preference.

male_mating_trait_loci=1:3: the loci specifying the male's mating trait.

competition_trait_loci=1:3: the loci specifying the ecological trait (used in fitness related to resource use).

hybrid_survival_loci=1:3: the loci specifying the probability of survival to adulthood.

survival_fitness_method:String="epistasis": the method used to calculate the probability of survival to adulthood.

per_reject_cost=0: the fitness loss of female per male rejected (due to search time, etc.). Can take values of 0 to 1.

sigma_disp=0.05: the standard deviation of the normal distribution determining how far offspring will disperse from their mothers.

track_spatial_data=false: if true, keeps track of bimodality, gene flow, variance, cline widths, and cline positions.

track_population_data=true: if true, keeps track of population size, hybridness, overlap, and hybrid zone width.

track_fitness=false: if true, keeps track of the average number of offspring for each phenotype.

track_mating_success=false: if true, keeps track of the average number of mates per male for each phenotype.

track_phenotypes=false: if true, keeps track of the number of individuals with each phenotype.

HZAM.plot_output_field — Method

plot_output_field(
    outcomes::Array{:Real},
    sim_params::Array{<:DataAnalysis.SimParams}
)

Create a heatmap of an output variable vs hybrid fitness and assortative mating.

Arguments

outcomes::Array{:Real}: the output from the simulation to be displayed.

sim_params::Array{<:DataAnalysis.SimParams}: the simulation parameters resulting in the outcomes.

HZAM.load_from_folder — Method

load_from_folder(dir::String)

Load the data from each simulation output file stored in the given directory into an array organized by hybrid fitness and assortative mating strength.

HZAM.load_from_csv — Method

load_from_csv(filepath::String)

Read the data from a CSV file and return vectors for hybrid fitness, assortative mating strength, and the output variable.

HZAM.convert_to_CSV — Method

convert_to_CSV(
    outcome_array::Array{<:DataAnalysis.OutputData},
    field_name::Symbol,
    output_folder::String
)

Convert an array of outcomes to a csv file for the given output field.

Arguments

outcome_array::Array{<:DataAnalysis.OutputData}: the array of output data from a set of simulations.

field_name::Symbol: the name of the output field of interest.

output_folder::String: the destination for the CSV.

HZAM.plot_fitnesses — Method

plot_fitnesses(fitnesses::Vector{<:Dict})

Produce a plot of fitnesses per phenotype over time.

Arguments

fitnesses::Vector{<:Dict}: number of offspring per phenotype each generation.

HZAM.plot_population_tracking_data — Method

plot_population_tracking_data(filepath::String)

Create plots of the population size, hybrid zone width, hybrid index, and population overlap vs time.

Arguments

filepath::String:: the filepath for the file containing the outcome of the simulation.

HZAM.summarize_gene_correlations — Method

summarize_gene_correlations(dir::String)

Create plots of the gene correlations between different traits vs SAM and whyb for different combinations of the same loci controlling different traits.

Arguments

dir::String: the folder where the outcomes of all the simulations are stored.

HZAM.plot_phenotypes — Method

plot_phenotypes(phenotypes; filepath="phenotype_frequencies.png")

Create plots of mating preference, mating cue, and hybrid survival trait phenotype frequencies over time.

Arguments

filepath="phenotype_frequencies.png": the destination filepath for the plot image.

Index

HZAM.DataAnalysis.OutputData

HZAM.DataAnalysis.PopulationTrackingData

HZAM.DataAnalysis.SimParams

HZAM.DataAnalysis.SpatialData

HZAM.Population.Location

HZAM.Population.PopulationData

HZAM.Population.Zone

HZAM.DataAnalysis.average_data_per_phenotype

HZAM.DataAnalysis.average_gene_data

HZAM.DataAnalysis.average_width

HZAM.DataAnalysis.calc_all_cline_widths_and_positions

HZAM.DataAnalysis.calc_all_gene_flow

HZAM.DataAnalysis.calc_all_trait_correlations

HZAM.DataAnalysis.calc_average_linkage_diseq

HZAM.DataAnalysis.calc_bimodality_in_range

HZAM.DataAnalysis.calc_bimodality_overall

HZAM.DataAnalysis.calc_chi_squared

HZAM.DataAnalysis.calc_length

HZAM.DataAnalysis.calc_linkage_diseq

HZAM.DataAnalysis.calc_overlap_in_range

HZAM.DataAnalysis.calc_overlap_overall

HZAM.DataAnalysis.calc_position

HZAM.DataAnalysis.calc_sigmoid_curve

HZAM.DataAnalysis.calc_sigmoid_curves

HZAM.DataAnalysis.calc_trait_correlation

HZAM.DataAnalysis.calc_traits_additive

HZAM.DataAnalysis.calc_variance

HZAM.DataAnalysis.calc_width

HZAM.DataAnalysis.count_phenotypes_at_loci

HZAM.DataAnalysis.find_fixed_alleles

HZAM.DataAnalysis.sigmoid

HZAM.DataAnalysis.sort_locations

HZAM.DataAnalysis.sort_y

HZAM.Mating.calc_match_strength

HZAM.Mating.choose_closest_male

HZAM.Mating.choose_closest_male_from_zone

HZAM.Mating.distance

HZAM.Mating.generate_offspring_genotype

HZAM.PlotData.count_genotypes

HZAM.PlotData.create_gene_plot

HZAM.PlotData.create_population_plot

HZAM.PlotData.get_expected

HZAM.PlotData.scale_curve

HZAM.PlotData.update_gene_plot

HZAM.PlotData.update_population_plot

HZAM.Population.assign_zone

HZAM.Population.calc_growth_rates

HZAM.Population.calc_ideal_densities

HZAM.Population.calc_ind_useResourceA

HZAM.Population.calc_ind_useResourceB

HZAM.Population.calc_real_densities

HZAM.Population.calc_squared_distances

HZAM.Population.calc_survival_fitness_epistasis

HZAM.Population.calc_survival_fitness_hetdisadvantage

HZAM.Population.genotype_mean

HZAM.Population.get_surrounding_zones

HZAM.Population.max_radius_squared

HZAM.convert_to_CSV

HZAM.load_from_csv

HZAM.load_from_folder

HZAM.plot_fitnesses

HZAM.plot_output_field

HZAM.plot_phenotypes

HZAM.plot_population_tracking_data

HZAM.run_HZAM_set

HZAM.run_one_HZAM_sim

HZAM.summarize_gene_correlations