DataAnalysis

HZAM.DataAnalysis — Module

Types and functions used for analyzing data produced by the simulation.

Constants

HZAM.DataAnalysis.initial_par — Constant

Parameters for fitting sigmoid curves to the clines.

HZAM.DataAnalysis.spaced_locations — Constant

Evenly spaced locations across the range (one-dimensional).

Types

HZAM.DataAnalysis.OutputData — Type

OutputData

An OutputData stores all of the key data from a simulation run.

fields

sim_params::SimParams: the parameters of the simulation.

genotypes::Vector{<:Matrix{<:Integer}}: the genotypes of the population at the end of the simulation.

x_locations::Vector{Float32}: the x coordinates of the population at the end of the simulation.

y_locations::Vector{Float32}: the y coordinates of the population at the end of the simulation.

hybrid_zone_width::Real: the cline width associated with the male mating trait.

population_overlap::Real: the proportion of the range occupied by males of both mating trait phenotypes.

bimodality::Real: the extent to which phenotypically pure individuals occur in the hybrid zone

population_tracking_data::Vector{PopulationTrackingData}: the population size, hybridity, overlap, and cline width over time.

HZAM.DataAnalysis.PopulationTrackingData — Type

PopulationTrackingData

A PopulationTrackingData stores the size, hybridity, overlap, and male mating trait cline width of the population.

As the simulation runs a vector of PopulationTrackingData is used to keep track of the key population data each generation.

Fields

population_size::Real: the total number of individuals.

hybridity::Real: the average hybridity in the population on the male mating trait loci.

overlap::Real: the proportion of the range containing both male mating trait phenotypes.

width::Real: the cline width for the male mating trait.

Constructors

- PopulationTrackingData(
    genotypes::Vector{<:Matrix{<:Integer}},
    locations::Vector,
    male_mating_trait_loci::Union{UnitRange{<:Integer},Vector{<:Integer}}
)

HZAM.DataAnalysis.SimParams — Type

SimParams

A SimParams stores the parameters used in a simulation.

Fields

intrinsic_R::Real: the intrinsic growth rate.

w_hyb::Real: the hybrid fitness.

S_AM::Real: the strength of assortative mating.

K_total::Integer: the carrying capacity of the environment.

max_generations::Integer: the number of generations the simulation ran for.

sigma_disp::Real: the standard deviation for dispersal distance.

total_loci::Integer: the number of loci in the genotypes.

female_mating_trait_loci: the loci controlling mating preference.

male_mating_trait_loci: the loci controlling mating cue.

hybrid_survival_loci: the loci controlling hybrid fitness.

per_reject_cost::Real: the loss in fitness incurred by a female after rejecting one male.

Functions

HZAM.DataAnalysis.average_width — Method

average_width(sigmoid_curves::Vector{<:Vector{<:Real}})

Compute the width of a cline given a series of sigmoid curves representing the cline along horizontal strips of the range.

HZAM.DataAnalysis.calc_bimodality_in_range — Method

calc_bimodality_in_range(
    sigmoid_curve::Vector{<:Real},
    locations_x::Vector{<:Real},
    hybrid_indices::Vector{<:Real},
    sigma_disp::Real
)

Compute the proportion of phenotypically pure individuals within half a dispersal distance of the cline midpoint.

Arguments

sigmoid_curve::Vector{<:Real}: the sigmoid curve fitting the cline.

locations_x::Vector{<:Real}: the x values of the locations.

hybrid_indices::Vector{<:Real}: the mean values of the genotypes over the functional loci only.

sigma_disp::Real: the standard deviation in the dispersal distance distribution.

HZAM.DataAnalysis.calc_bimodality_overall — Method

Compute the bimodality across the entire range (see above for further description).

calc_bimodality_overall(
    sigmoid_curves::Vector{<:Vector{<:Real}},
    sorted_indices::Vector{<:Vector{<:Integer}},
    locations_x::Vector{<:Real},
    hybrid_indices::Vector{<:Real},
    sigma_disp::Real
)

Arguments

sigmoid_curves::Vector{<:Real}: the sigmoid curves for each horizontal strip of the range.

sorted_indices::Vector{<:Vector{<:Integer}}: the indices of the locations sorted into bins bins corresponding to non-overlapping ranges of the y values.

locations_x::Vector{<:Real}: the x values of the locations.

hybrid_indices::Vector{<:Real}: the mean values of the genotypes over the functional loci only.

sigma_disp::Real: the standard deviation in the dispersal distance distribution.

HZAM.DataAnalysis.calc_sigmoid_curve — Method

calc_sigmoid_curve(locations_x::Vector{<:Real}, hybrid_indices::Vector{<:Real})

Compute a sigmoid curve that models hybrid index vs location on the x axis.

Return the output of the sigmoid curve function on 1000 evenly spaced locations.

HZAM.DataAnalysis.calc_sigmoid_curves — Method

calc_sigmoid_curves(x_locations::Vector{Float32}, y_locations::Vector{Float32}, hybrid_indices::Vector{<:Real})

Divide the range into ten horizontal strips and compute a sigmoid curve fitting the hybrid indices along each strip.

HZAM.DataAnalysis.calc_traits_additive — Method

calc_traits_additive(
    genotypes::Vector{<:Matrix{<:Integer}},
    loci::Union{UnitRange{<:Integer},Vector{<:Integer}}
)::Vector{Float32}

Compute the mean values of the genotypes passed to it at the given loci. Used to determine trait values in an additive way.

HZAM.DataAnalysis.calc_width — Method

calc_width(sigmoid_curve::Vector{<:Real})

Compute the width of a cline given a sigmoid curve.

The cline width is defined as the distance between where the sigmoid curve passes through 0.1 and 0.9.

HZAM.DataAnalysis.sigmoid — Method

sigmoid(x::Vector{<:Real}, p::Vector{<:Real})

Compute the y value for a sigmoid given the x value, centre, and maximum slope.

Arguments

x::Real: the value along the x axis where the sigmoid is to be calculated.

p::Vector{<:Real}: the fit parameters where p[1] is the cline centre and p[2] is

the maximum slope.

HZAM.DataAnalysis.sort_locations — Method

sort_locations(A::AbstractArray{<:Real}, bin_size::Real)

Sort the given list of numbers into bins of a given size spanning 0 to 1.

HZAM.DataAnalysis.sort_y — Method

sort_y(y_locations::Vector{<:Real})

Sort the y coordinate indices into 10 vectors [0, 0.1), [0.1, 0.2), etc.

Example

jldoctest sort_y([0.01, 0.5, 0.24, 0.9]) 10-element Vector{<:Vector{Int64}}: [1] [] [3] [] [] [2] [] [] [] [4]