Mating

HZAM.Mating — Module

Functions involved in mate selection and generating the offspring genotype.

Functions

HZAM.Mating.calc_match_strength — Method

calc_match_strength(
    female_genotype::Matrix{<:Integer}, 
    male_genotype::Matrix{<:Integer}, 
    pref_SD::Float64, 
    female_mating_trait_loci, 
    male_mating_trait_loci
)

Compare the male's mating trait with the female mating trait to determine the match strength along a Gaussian acceptance curve.

The acceptance curve is bounded between 0 and 1.

Arguments

female_genotype::Matrix{<:Integer}: the genotype of the focal female.

male_genotype::Matrix{<:Integer}: the genotype of the focal male.

pref_SD::Float64: width of the Gaussian acceptance curve.

female_mating_trait_loci: the loci contributing to the female's mate preference.

male_mating_trait_loci: the loci contributing to the male's trait.

HZAM.Mating.choose_closest_male — Method

choose_closest_male(
    zones::Matrix{Zone}, 
    zone_indices::Vector{CartesianIndex{2}}, 
    elig_M::Dict{CartesianIndex,<:Vector{<:Integer}}, 
    location_mother::Location, 
    neighbourhood_size::Float32
)

Find the index of the closest male within an area covering multiple zones from a list of eligible males.

Return both the male index and its zone index. Return -1 for both values if no eligible male is found.

Arguments

zones::Matrix{Zone}: the matrix storing the population data for each zone.

zone_indices::Vector{CartesianIndex{2}}: the indices of the zones to be checked for the closest male.

elig_M::Dict{CartesianIndex,<:Vector{<:Integer}}: the indices of all the eligible males in each zone.

location_mother::Location: the location from which the males' distances are computed.

neighbourhood_size::Float32: the distance cutoff for the search.

HZAM.Mating.choose_closest_male_from_zone — Method

choose_closest_male_from_zone(
    elig_M::Vector{<:Integer}, 
    locations_M::Vector{Location}, 
    location_mother::Location
)

Find the index of the closest male within a single zone from a list of eligible males.

Arguments

elig_M::Vector{<:Integer}: the indices of all the eligible males.

locations_M::Vector{Location}: the locations of all the males in the zone.

location_mother::Location: the location from which the males' distances are computed.

HZAM.Mating.distance — Method

distance(location1::Location, location2::Location)

Compute the distance between location1 and location2.

HZAM.Mating.generate_offspring_genotype — Method

generate_offspring_genotype(
    mother_genotype::Matrix{<:Integer}, 
    father_genotype::Matrix{<:Integer}
)::Matrix{<:Integer}

Generate the offspring genotype from the parent genotypes.

For each locus (column), first row for allele from mother, second row for allele from father. The number of loci in each genotype must be equal.

Example

julia> generate_offspring_genotype([0 0 0; 0 0 0], [1 1 1; 1 1 1])
2×3 Matrix:
 0  0  0
 1  1  1