PlotData

HZAM.PlotData — Module

Functions for plotting data while the simulation is running.

Functions

HZAM.PlotData.count_genotypes — Method

function count_genotypes(
    range::Union{UnitRange{<:Integer},Vector{<:Integer}},
    genotypes::Vector{<:Matrix{<:Integer}}
)

Count the number of individuals with each genotype for a given loci range.

HZAM.PlotData.create_gene_plot — Method

create_gene_plot(
    genotypes::Vector{<:Matrix{<:Integer}},
    loci::NamedTuple,
    save_plot::Bool
)

Initialize the plot of the phenotype frequencies for each loci range.

Arguments

genotypes::Vector{<:Matrix{<:Integer}}: the genotypes of every individual.

loci::NamedTuple: the loci range for each trait.

save_plot::Bool: true if the plot is to be saved to a file.

HZAM.PlotData.create_population_plot — Method

create_population_plot(
    hybrid_indices_functional::Vector{<:Real},
    locations::Vector,
    save_plot::Bool
)

Initialize the plot of locations and hybrid indices.

Arguments

hybrid_indices_functional::Vector: list of the hybrid index (value between 0 and 1) of

every individual.

locations::Vector: the location of every individual

(must be in the same order as the hybrid indices).

generation::Integer: the number of elapsed generations.

save_plot::Bool: true if the plot is to be saved to a PNG file.

HZAM.PlotData.get_expected — Method

get_expected(x::Real, n::Integer)

Compute the expected number of individuals with a given trait hybrid index if the trait is controlled by 4 loci.

Arguments

x::Real: the hybrid index for the trait of interest.

n::Integer: the total population size.

HZAM.PlotData.update_gene_plot — Method

update_gene_plot(
    genotypes::Vector{<:Matrix{<:Integer}},
    loci::NamedTuple,
    generation::Integer,
    save_plot::Bool
)

Update the plot of the phenotype frequencies for each loci range.

Arguments

genotypes::Vector{<:Matrix{<:Integer}}: the genotypes of every individual.

loci::NamedTuple: the loci range for each trait.

generation::Integer: the generation # the simulation is on.

save_plot::Bool: true if the plot is to be saved to a file.

HZAM.PlotData.update_population_plot — Method

update_population_plot(
    hybrid_indices_functional::Vector,
    locations::Vector,
    generation::Integer,
    save_plot::Bool
)

Update the existing plot with new locations and hybrid indices.

Arguments

hybrid_indices_functional::Vector: list of the hybrid index (value between 0 and 1) of

every individual.

locations::Vector: the location of every individual (must be in the same order as the

hybrid indices).

generation::Integer: the number of elapsed generations.

save_plot::Bool: true if the plot is to be saved to a PNG file.