DataAnalysis

HZAM.DataAnalysis — Module

Types and functions used for analyzing data produced by the simulation.

Constants

HZAM.DataAnalysis.initial_par — Constant

Parameters for fitting sigmoid curves to the clines.

HZAM.DataAnalysis.spaced_locations — Constant

Evenly spaced locations across the range (one-dimensional).

Types

HZAM.DataAnalysis.OutputData — Type

OutputData

An OutputData stores all of the key data from a simulation run.

Fields

sim_params::SimParams: the parameters of the simulation.

population_data: the genotypes and locations for all individuals in the simulation.

hybrid_zone_width::Vector{<:Real}: the cline width associated with the male mating trait.

population_overlap::Vector{<:Real}: the proportion of the range occupied by males of both mating trait phenotypes.

bimodality::Real: the extent to which phenotypically pure individuals occur in the hybrid zone

population_tracking_data::Vector{PopulationTrackingData}: the population size, hybridity, overlap, and cline width over time.

overlap::Real: the proportion of the range where both species occur.

HZAM.DataAnalysis.SimParams — Type

SimParams

A SimParams stores the parameters used in a simulation.

Fields

intrinsic_R::Real: the intrinsic growth rate.

w_hyb::Real: the hybrid fitness.

S_AM::Real: the strength of assortative mating.

K_total::Integer: the carrying capacity of the environment.

max_generations::Integer: the number of generations the simulation ran for.

sigma_disp::Real: the standard deviation for dispersal distance.

total_loci::Integer: the number of loci in the genotypes.

female_mating_trait_loci: the loci controlling mating preference.

male_mating_trait_loci: the loci controlling mating cue.

hybrid_survival_loci: the loci controlling hybrid fitness.

per_reject_cost::Real: the loss in fitness incurred by a female after rejecting one male.

Functions

HZAM.DataAnalysis.calc_bimodality — Method

calc_bimodality(
	pd,
	loci::Union{UnitRange{<:Integer}, Vector{<:Integer}},
)

Compute the bimodality of the hybrid zone by calculating the bimodality along 9 transects.

Arguments

pd: the PopulationData containing the genotypes, x coordinates, and y coordinates of

all individuals.

loci::Union{UnitRange{<:Integer}, Vector{<:Integer}}: the loci specifying the trait of

interest.

HZAM.DataAnalysis.calc_bimodality_on_transect — Method

calc_bimodality_on_transect(
	pd,
	sigmoid_curve::Vector{<:Real},
	y_coord::Real,
	loci::Union{UnitRange{<:Integer}, Vector{<:Integer}}
)

Compute the bimodality along a transect by locating the midpoint and calculating the density of phenotypically pure individuals there.

Arguments

pd: the PopulationData containing the genotypes, x coordinates, and y coordinates of

all individuals.

sigmoid_curve::Vector{<:Real}: a sigmoid curve fit to the cline along the transect.

y_coord::Real: the y coordinate of the transect.

loci::Union{UnitRange{<:Integer}, Vector{<:Integer}}: the loci specifying the trait of

interest.

HZAM.DataAnalysis.calc_cline_width — Method

calc_cline_width(
	pd,
	loci::Union{UnitRange{<:Integer}, Vector{<:Integer}},
	y_coords::Union{StepRangeLen{<:Real}, Vector{<:Real}},
)

Compute the cline width by averaging the width along transects at the given y coordinates.

Arguments

pd: the PopulationData containing the genotypes, x coordinates, and y coordinates of

all individuals.

y_coords: the y coordinates of the transects used to compute the width.

loci::Union{UnitRange{<:Integer}, Vector{<:Integer}}: the loci specifying the trait of

interest.

HZAM.DataAnalysis.calc_sigmoid_curve — Method

calc_sigmoid_curve(locations_x::Vector{<:Real}, hybrid_indices::Vector{<:Real})

Compute a sigmoid curve that models hybrid index vs location on the x axis.

Return the output of the sigmoid curve function on 1000 evenly spaced locations.

HZAM.DataAnalysis.calc_traits_additive — Method

calc_traits_additive(
	genotypes::Vector{<:Matrix{<:Integer}},
	loci::Union{UnitRange{<:Integer},Vector{<:Integer}}
)::Vector{Float32}

Compute the mean values of the genotypes passed to it at the given loci. Used to determine trait values in an additive way.

HZAM.DataAnalysis.calc_transect — Method

calc_transect(
	genotypes::Vector{<:Matrix{<:Integer}}, 
	x_locations::Vector{<:Real}, 
	y_locations::Vector{<:Real}, 
	loci::Union{UnitRange{<:Integer},Vector{<:Integer}}, 
	sigma_comp::Real, 
	y_coord::Real
)

Compute the cline width along a horizontal transect.

Arguments

genotypes::Vector{<:Matrix{<:Integer}} : the genotypes of every individual in the sim.

x_locations::Vector{<:Real} : the x coordinate of every individual in the sim.

y_locations::Vector{<:Real} : the y coordinate of every individual in the sim.

loci::Union{UnitRange{<:Integer},Vector{<:Integer}} : the focal loci.

sigma_comp::Real : the standard deviation used for the density calculation.

y_coord::Real : the y coordinate of the transect along which cline width is measured.

HZAM.DataAnalysis.calc_transects — Method

calc_transects(
	pd,
	loci::Union{UnitRange{<:Integer}, Vector{<:Integer}},
	y_coords::Union{StepRangeLen{<:Real}, Vector{<:Real}},
)

Compute a sigmoidal cline model for the specified loci at horizontal transects with the specified y coordinates. Returns the weighted average hybrid index at evenly spaced locations along the transect, the sigmoidal curves modeling the transects, and the cline widths.

Arguments

pd: the PopulationData containing the genotypes, x coordinates, and y coordinates of

all individuals.

y_coords: the y coordinates of the transects.

loci::Union{UnitRange{<:Integer}, Vector{<:Integer}}: the loci specifying the trait of

interest.

HZAM.DataAnalysis.count_phenotypes_at_loci — Method

function count_phenotypes_at_loci(
	genotypes::Vector{<:Matrix{<:Integer}},
	loci::Union{UnitRange{<:Integer},Vector{<:Integer}}
)

Return the number of individuals with each phenotype for the given loci range.

Arguments

genotypes::Vector{<:Matrix{<:Integer}}: the genotypes of every individual

loci::Union{UnitRange{<:Integer},Vector{<:Integer}}: the loci range over which the phenotypes are calculated.

HZAM.DataAnalysis.sigmoid — Method

sigmoid(x::Vector{<:Real}, p::Vector{<:Real})

Compute the y value for a sigmoid given the x value, centre, and maximum slope.

Arguments

x::Real: the value along the x axis where the sigmoid is to be calculated.

p::Vector{<:Real}: the fit parameters where p[1] is the cline centre and p[2] is

the maximum slope.