HZAM

Index

HZAM.Population

HZAM.Population.NUM_ZONES

HZAM.Population.PopulationData

HZAM.Population.Zone

HZAM.Population.assign_zone

HZAM.Population.calc_growth_rates

HZAM.Population.calc_ideal_densities

HZAM.Population.calc_real_densities

HZAM.Population.calc_species_overlap

HZAM.Population.calc_squared_distances

HZAM.Population.calc_survival_fitness_epistasis

HZAM.Population.calc_survival_fitness_hetdisadvantage

HZAM.Population.genotype_mean

HZAM.Population.get_surrounding_zones

HZAM.Population.max_radius_squared

HZAM.Population.new_location

HZAM.Population.new_location

HZAM.Mating

HZAM.Mating.calc_match_strength

HZAM.Mating.calc_match_strength_multivariate

HZAM.Mating.choose_closest_male

HZAM.Mating.choose_closest_male_from_zone

HZAM.Mating.distance

HZAM.Mating.generate_offspring_genotype

HZAM.DataAnalysis

HZAM.DataAnalysis.initial_par

HZAM.DataAnalysis.spaced_locations

HZAM.DataAnalysis.OutputData

HZAM.DataAnalysis.SimParams

HZAM.DataAnalysis.calc_bimodality

HZAM.DataAnalysis.calc_bimodality_on_transect

HZAM.DataAnalysis.calc_cline_width

HZAM.DataAnalysis.calc_sigmoid_curve

HZAM.DataAnalysis.calc_traits_additive

HZAM.DataAnalysis.calc_transect

HZAM.DataAnalysis.calc_transects

HZAM.DataAnalysis.count_phenotypes_at_loci

HZAM.DataAnalysis.sigmoid

HZAM.PlotData

HZAM.PlotData.count_genotypes

HZAM.PlotData.create_gene_plot

HZAM.PlotData.create_population_plot

HZAM.PlotData.get_expected

HZAM.PlotData.update_gene_plot

HZAM.PlotData.update_population_plot

Functions

HZAM.run_one_HZAM_sim — Method

run_one_HZAM_sim(
	w_hyb::Real, 
	S_AM::Real,
	intrinsic_R::Real; 
	<keyword arguments>
)

Run a single HZAM simulation.

Arguments

w_hyb::Real: the hybrid fitness.

S_AM::Real: the strength of assortative mating.

intrinsic_R::Real: the intrinsic growth rate.

K_total::Integer=20000: the carrying capacity of the environment.

max_generations::Integer=1000: the number of generations that the simulation will run for.

total_loci::Integer=6: the total number of loci in the genome.

female_mating_trait_loci=1:3: the loci specifying the female's mate preference.

male_mating_trait_loci=1:3: the loci specifying the male's mating trait.

hybrid_survival_loci=1:3: the loci specifying the probability of survival to adulthood.

survival_fitness_method:String="epistasis": the method used to calculate the probability of survival to adulthood.

per_reject_cost=0: the fitness loss of female per male rejected (due to search time, etc.). Can take values of 0 to 1.

sigma_disp::Float32=0.05f0: the standard deviation of the normal distribution determining how far offspring will disperse from their mothers.

sigma_comp::Float32=0.01f0: the standard deviation for the normal curve used in calculating local density.

do_plot=true: the program will display a plot of the locations and hybrid indices of every individual while the simulation is running if this is true.

plot_int=10: the interval (measured in generations) between updating the plot.

gene_plot=false: if true, generates phenotype plots.

save_plot=false: if true, saves each plot to a PNG file.

run_name="temp": the name of the output file

cline_width_loci: either "mmt" or "fmt", sets the loci used to determine simulation outcomes.

mating_preference_type: either "additive" or "multivariate", determines how the preference loci affect match strength.

HZAM.categorize — Method

categorize(outcome)

Classify the simulation outcome into one of six outcome types based on the cline width, bimodality, and population overlap.

HZAM.extract_all — Method

extract_all(source_dir::String, output_dir::String; set_numbers::Union{UnitRange{<:Integer}, Vector{<:Integer}} = 1:9, categorization_loci = "mmt")

Condense all of the simulation results data into a format easier to work with.

Arguments

source_dir::String: the directory storing the simulation outcomes

output_dir::String: the output directory to be created

set_numbers::Union{UnitRange{<:Integer}, Vector{<:Integer}} = 1:9: the simulation setups used

categorization_loci = "mmt": the loci used to categorized simulation outcomes (either "fmt" or "mmt")

HZAM.extract_data — Method

extract_data(source_folder::String, source_setup::String, output_folder::String, categorization_loci::String)

Load all of the simulation outcomes in a folder and condense the data by removing anything not used in the paper.

Arguments

source_folder::String: the folder where the simulation outcomes are stored

source_setup::String: the simulation setup of interest (i.e. "no_pleiotropy)

output_folder::String: where the output is to be stored

categorization_loci::String: whether mmt or fmt is used for categorizing outcomes

HZAM.first_to_three — Method

first_to_three(source_dir::String, output_dir::String; set_numbers::Union{UnitRange{<:Integer}, Vector{<:Integer}} = 1:9, categorization_loci = "mmt")

Check if any parameter setups have not resulted in three of the same outcome type and run another round of simulations for those parameter setups.

Arguments

source_dir::String: the directory storing the simulation outcomes

output_dir::String: the output directory to be created

set_numbers::Union{UnitRange{<:Integer}, Vector{<:Integer}} = 1:9: the simulation setups used

categorization_loci = "mmt": the loci used to categorized simulation outcomes (either "fmt" or "mmt")

HZAM.is_blended — Method

is_blended(outcome::HZAM.DataAnalysis.OutputData)::Bool

Determine if a simulation outcome is blended.

HZAM.is_one_species_extinct — Method

is_one_species_extinct(outcome::HZAM.DataAnalysis.OutputData; categorization_loci = "mmt")::Bool

Determine if the simulation outcome shows the extinction of one population.

HZAM.load_from_folder — Method

load_from_folder(dir::String)

Load the data from each simulation output file stored in the given directory into an array organized by hybrid fitness and assortative mating strength.

HZAM.mean — Method

Compute the mean value of a vector.

HZAM.plot_fitnesses — Method

plot_fitnesses(fitnesses::Vector{<:Dict})

Produce a plot of fitnesses per phenotype over time.

Arguments

fitnesses::Vector{<:Dict}: number of offspring per phenotype each generation.

HZAM.plot_population_tracking_data — Method

plot_population_tracking_data(filepath::String)

Create plots of the population size, hybrid zone width, hybrid index, and population overlap vs time.

Arguments

filepath::String:: the filepath for the file containing the outcome of the simulation.

HZAM.read_directory — Method

read_directory(dir::String, set::String)

Load the data from each simulation output style in each run folder in the given directory for the given simulation set (i.e. "full_pleiotropy").

HZAM.run_HZAM_set — Method

run_HZAM_set(
	set_name::String,
	total_loci::Int,
	female_mating_trait_loci::Union{UnitRange{<:Integer}, Vector{<:Integer}},
	male_mating_trait_loci::Union{UnitRange{<:Integer}, Vector{<:Integer}},
	hybrid_survival_loci::Union{UnitRange{<:Integer}, Vector{<:Integer}},
	per_reject_cost::Real,
	w_hyb_set::Union{UnitRange{<:Real}, Vector{<:Real}},
	S_AM_set::Union{UnitRange{<:Real}, Vector{<:Real}},
	dir::String;
	intrinsic_R::Real = 1.1, K_total::Int = 20000, max_generations::Int = 1500,
	survival_fitness_method::String = "epistasis", sigma_disp = 0.03,
)

Run the simulation for the given combinations of hybrid fitness and assortative mating strength and store the outcome of each simulation in a JLD2 file.

Arguments

set_name::String: the name assigned to the set of simulations.

intrinsic_R::Real: the intrinsic growth rate.

K_total::Integer=20000: the carrying capacity of the environment.

max_generations::Integer=1500: the number of generations that the simulation will run.

total_loci::Integer=6: the total number of loci in the genome.

female_mating_trait_loci=1:3: the loci specifying the female's mate preference.

male_mating_trait_loci=1:3: the loci specifying the male's mating trait.

hybrid_survival_loci=1:3: the loci specifying the probability of survival to adulthood.

survival_fitness_method:String="epistasis": the method used to calculate the probability

of survival to adulthood.

per_reject_cost=0: the fitness loss of female per male rejected (due to search time,

etc.). Can take values of 0 to 1.

sigma_disp=0.05: the standard deviation of the normal distribution determining how far

offspring will disperse from their mothers.

HZAM.run_HZAM_sets_complete_nine_loci — Method

run_HZAM_sets_complete_nine_loci(;
	set_numbers::Union{UnitRange{<:Integer}, Vector{<:Integer}} = 1:9,
	w_hyb_set_of_run::Union{UnitRange{<:Real}, Vector{<:Real}} = w_hyb_set,
	S_AM_set_of_run::Union{UnitRange{<:Real}, Vector{<:Real}} = S_AM_set,
	max_generations::Integer = 1500,
	K_total::Integer = 20000,
)

Run the main set of simulations with nine loci per trait and save the outcomes to results_folder.

Arguments

set_numbers::Union{UnitRange{<:Integer}, Vector{<:Integer}}: which sets to run (the sets

are full pleiotropy, no pleiotropy, magic cue, magic preference, etc.).

w_hyb_set_of_run::Union{UnitRange{<:Real}, Vector{<:Real}}: which w_hyb values to use.

S_AM_set_of_run::Union{UnitRange{<:Real}, Vector{<:Real}}: which S_AM values to use.

max_generations::Integer: how many generations to run the simulations for.

K_total::Integer: the carrying capacity of the simulated range.

HZAM.run_HZAM_sets_complete_one_locus — Method

run_HZAM_sets_complete_one_locus(;
	set_numbers::Union{UnitRange{<:Integer}, Vector{<:Integer}} = 1:9,
	w_hyb_set_of_run::Union{UnitRange{<:Real}, Vector{<:Real}} = w_hyb_set,
	S_AM_set_of_run::Union{UnitRange{<:Real}, Vector{<:Real}} = S_AM_set,
	max_generations::Integer = 1500,
	K_total::Integer = 20000,
)

Run the main set of simulations with one locus per trait and save the outcomes to results_folder.

Arguments

set_numbers::Union{UnitRange{<:Integer}, Vector{<:Integer}}: which sets to run (the sets

are full pleiotropy, no pleiotropy, magic cue, magic preference, etc.).

w_hyb_set_of_run::Union{UnitRange{<:Real}, Vector{<:Real}}: which w_hyb values to use.

S_AM_set_of_run::Union{UnitRange{<:Real}, Vector{<:Real}}: which S_AM values to use.

max_generations::Integer: how many generations to run the simulations for.

K_total::Integer: the carrying capacity of the simulated range.

HZAM.run_HZAM_sets_complete_three_loci — Method

run_HZAM_sets_complete_three_loci(;
	set_numbers::Union{UnitRange{<:Integer}, Vector{<:Integer}} = 1:9,
	w_hyb_set_of_run::Union{UnitRange{<:Real}, Vector{<:Real}} = w_hyb_set,
	S_AM_set_of_run::Union{UnitRange{<:Real}, Vector{<:Real}} = S_AM_set,
	max_generations::Integer = 1500,
	K_total::Integer = 20000,
)

Run the main set of simulations with three loci per trait and save the outcomes to results_folder.

Arguments

set_numbers::Union{UnitRange{<:Integer}, Vector{<:Integer}}: which sets to run (the sets

are full pleiotropy, no pleiotropy, magic cue, magic preference, etc.).

w_hyb_set_of_run::Union{UnitRange{<:Real}, Vector{<:Real}}: which w_hyb values to use.

S_AM_set_of_run::Union{UnitRange{<:Real}, Vector{<:Real}}: which S_AM values to use.

max_generations::Integer: how many generations to run the simulations for.

K_total::Integer: the carrying capacity of the simulated range.

HZAM.set_results_folder — Method

set_results_folder(dir::String)

Set the output parent directory for all methods in summarize_data.jl.

HZAM.summarize_gene_correlations — Method

summarize_gene_correlations(source_dir::String; filename="gene_correlations")

Create plots of the gene correlations between different traits vs SAM and whyb for different combinations of the same loci controlling different traits.

sourcedir should point to a folder containing magicpreference.jld2, magiccue.jld2, searchcost.jld2, and no_magic.jld2. Each file should contain an outcome array from a simulation set.

Arguments

source_dir::String: the folder where the outcomes of all the simulations are stored.

file_name="gene_correlations: the name of the image file for the plot.