Population

HZAM.Population — Module

Types and functions used for managing population data (locations, genotypes, and growth rates).

Constants

HZAM.Population.NUM_ZONES — Constant

The width of the (square) grid of zones that divides the population into more manageable chunks. Default is 10x10.

Types

HZAM.Population.PopulationData — Type

PopulationData

A PopulationData stores the genotypes, locations, and growth rates for all individuals in the simulation. The data is subdivided by 'zone' into a matrix for calculation efficiency.

Fields

population::Matrix{Zone}: all of the zones (containing the genotypes and locations) in the simulation.

growth_rates_F::Matrix{Vector{Float64}}: the growth rates of every female for each zone.

ideal_densities::Dict{Tuple{Float32, Float32}, Real}: the expected local densities, based on geographically even distribution of individuals at carrying capacity.

Constructors

- function PopulationData(
	K_total::Integer,
	total_loci::Integer,
	intrinsic_R::Real,
	sigma_comp::Real
)
- PopulationData(
	genotypes_daughters::Matrix{Vector{Matrix{Int8}}},
	genotypes_sons::Matrix{Vector{Matrix{Int8}}},
	x_locations_daughters::Matrix{Vector{Float32}},
	y_locations_daughters::Matrix{Vector{Float32}},
	x_locations_sons::Matrix{Vector{Float32}},
	y_locations_sons::Matrix{Vector{Float32}},
	sigma_comp::Real,
	intrinsic_R::Real,
	ideal_densities::Dict{Tuple{Float32, Float32}, Real}
)

Details on behaviour of different constructors

The first constructor sets up the initial population for the simulation.

The second constructor creates a new PopulationData using the genotypes and locations of the offspring.

HZAM.Population.Zone — Type

Zone

A Zone stores the genotypes and locations for the individuals contained within a subset of the range.

Fields

genotypes_F::Vector{Matrix{Int8}}: the female genotypes. rows are alleles (row 1 from mother, row 2 from father) and columns are loci.

genotypes_M::Vector{Matrix{Int8}}: the male genotypes.

x_locations_F::Vector{Float32}: the x coordinates of the females

y_locations_F::Vector{Float32}: the y coordinates of the females

x_locations_M::Vector{Float32}: the x coordinates of the males

y_locations_M::Vector{Float32}: the y coordinates of the males

Constructors

- Zone(
	starting_N::Integer,
	total_loci::Integer,
	x_location::Float32,
	y_location::Float32,
	size::Float32,
	species::Real
)
- Zone(
	genotypes_F::Vector{Matrix{Int8}},
	genotypes_M::Vector{Matrix{Int8}},
	x_locations_F::Vector{Float32},
	y_locations_F::Vector{Float32},
	x_locations_M::Vector{Float32},
	y_locations_M::Vector{Float32}
)

Details on behaviour of different constructors

The first constructor sets up the initial locations and genotypes of every individual in the zone.

The second constructor creates a new Zone using the genotypes and locations of the offspring.

Functions

HZAM.Population.assign_zone — Method

assign_zone(x::Float32, y::Float32)

Determine which zone a location falls in.

HZAM.Population.calc_growth_rates — Method

calc_growth_rates(
	population::Matrix{Zone},
	zone_index::CartesianIndex,
	sigma_comp::Real,
	intrinsic_R::Real,
	ideal_densities::Dict{Tuple{Float32, Float32}, Real}
)

Compute the female growth rates assuming no ecological difference.

Arguments

population::Matrix{Zone}: the matrix of zones storing all the locations, and genotypes in the simulation.

zone_index::CartesianIndex: the index of the focal zone.

sigma_comp::Real: the standard deviation for the normal curve used in calculating local density.

intrinsic_R::Real: the intrinsic growth rate.

ideal_densities::Dict{Tuple{Float32, Float32}, Real}: the expected local densities, based on geographically even distribution of individuals at carrying capacity.

HZAM.Population.calc_ideal_densities — Method

calc_ideal_densities(
	K_total::Integer,
	sigma_comp::Real,
	x_locations_F::Vector{Float32},
	y_locations_F::Vector{Float32},
	max_dist::Real
)

Compute expected local densities, based on geographically even distribution of individuals at carrying capacity.

Densities are calculated using the following integral:

$K_{total}\int\limits_0^{2\pi}\int\limits_0^{0.03} \exp(-\frac{r^2}{2σ_{comp}^2})rdrdθ$

where the density function is equal to 0 outside of the range limits.

Argument

K_total::Integer: the carrying capacity

sigma_comp::Real: the standard deviation for the normal curve used in calculating local density.

x_locations_F::Vector{Float32}: the x coordinates where the ideal density is to be computed.

y_locations_F::Vector{Float32}: the y coordinates where the ideal density is to be computed.

max_dist::Real: the cutoff for the furthest away an individual can be and affect the density calculation. Should be 3x sigma_comp.

HZAM.Population.calc_real_densities — Method

calc_real_densities(
	population::Matrix{Zone},
	zone_index::CartesianIndex,
	x_locations_F::Vector{Float32},
	y_locations_F::Vector{Float32},
	max_dist::Real,
	sigma_comp::Real;
	species=-1,
	loci=[-1]
)

Compute the population density at each female's location.

Arguments

population::Matrix{Zone}: the matrix of zones storing all the locations, and genotypes in the simulation.

zone_index::CartesianIndex: the index of the focal zone.

x_locations_F::Vector{Float32}: the x coordinates of every female in the focal zone.

y_locations_F::Vector{Float32}: the y coordinates of every female in the focal zone.

max_dist::Real: the distance cutoff for the density calculation.

sigma_comp::Real: the standard deviation used in the density calculation.

HZAM.Population.calc_species_overlap — Method

calc_species_overlap(
	population::Matrix{Zone},
	max_dist::Real,
	sigma_comp::Real,
	loci::Union{UnitRange{<:Integer},Vector{<:Integer}}
)

Compute the total overlap between the two species.

Arguments

population::Matrix{Zone}: the genotypes and locations of all individuals.

max_dist::Real: the distance cutoff for the density calculation.

sigma_comp::Real: the standard deviation used for the density calculation.

loci::Union{UnitRange{<:Integer},Vector{<:Integer}}: the loci used to determine species.

HZAM.Population.calc_squared_distances — Method

function calc_squared_distances(
	x_locations::Vector{Float32},
	y_locations::Vector{Float32},
	x::Float32,
	y::Float32,
	cutoff::Real
)

Compute the squared distances from a set of locations to a single location. Discards distances above the given cutoff.

Example

julia> calc_squared_distances([0.4f0, 0f0, 0.5f0], [0.4f0, 0f0, 0.5f0],0.1)
2-element Vector{Float32}:
 0.0
 0.019999998

HZAM.Population.calc_survival_fitness_epistasis — Function

calc_survival_fitness_epistasis(
	genotype::Matrix{<:Integer}, 
	hybrid_survival_loci::Union{UnitRange{<:Integer},Vector{<:Integer}}, 
	w_hyb::Real, 
	beta::Real=1
)

Compute the survival fitness of an individual according to epistasis, with the beta parameter set to one as a default.

Arguments

genotype:Matrix{<:Integer}: the individual's genotype.

hybrid_survival_loci::Union{UnitRange{<:Integer},Vector{<:Integer}}: the loci responsible for heterozygote disadvantage in computing the survival fitness.

w_hyb::Real: the simulation's hybrid fitness value.

beta::Real

HZAM.Population.calc_survival_fitness_hetdisadvantage — Method

calc_survival_fitness_hetdisadvantage(genotype::Matrix{<:Integer}, w_hyb::Real)

Compute the survival fitness of an individual according to heterozygosity.

Arguments

genotype:Matrix{<:Integer}: the individual's genotype.

w_hyb::Real: the simulation's hybrid fitness value.

HZAM.Population.genotype_mean — Method

genotype_mean(genotype::Matrix{<:Integer}, loci::Union{UnitRange{<:Integer},Vector{<:Integer}})

Compute the mean value of the genotype across the list of loci given.

Example

julia> genotype_mean([0 1 0; 0 1 0], [1,3])
0.0

HZAM.Population.get_surrounding_zones — Method

get_surrounding_zones(zone_index::CartesianIndex)

Return a list of the indices of the neighbouring zones and the given zone index.

HZAM.Population.max_radius_squared — Method

max_radius_squared(x::Float32, y::Float32, t::Real, max_dist::Real)

Compute the distance from a point along an angle to the limit of the range (defined by x∈[0,1), y∈[0,1)).

The distance gets cut off at max_dist. Used in calculating the ideal densities.

Arguments

x::Float32: the starting x coordinate

y::Float32: the starting y coordinate

t::Real: the angle.

max_dist: the maximum distance.

HZAM.Population.new_location — Method

new_location(x::Float32, y::Float32, sigma_disp::Real)::Tuple{Float32,Float32}

Compute the location after dispersal of an offspring.

Arguments

x::Float32: the x coordinate of the mother

y::Float32: the y coordinate of the mother

sigma_disp::Real: the standard deviation in the dispersal distance

HZAM.Population.new_location — Method

new_location(
	starting_range_x::Vector{Float32},
	starting_range_y::Vector{Float32}
)::Tuple{Float32,Float32}

Randomly generate a location within the given range.

Arguments

starting_range_x::Vector{Float32}: the x limits of the range

starting_range_y::Vector{Float32}: the y limits of the range