DataAnalysis

HZAM.DataAnalysis — Module

Types and functions used for analyzing data produced by the simulation.

Constants

HZAM.DataAnalysis.initial_par — Constant

Parameters for fitting sigmoid curves to the clines.

HZAM.DataAnalysis.spaced_locations — Constant

Evenly spaced locations across the range (one-dimensional).

Types

HZAM.DataAnalysis.OutputData — Type

OutputData

An OutputData stores all of the key data from a simulation run.

fields

sim_params::SimParams: the parameters of the simulation.

genotypes::Vector{<:Matrix{<:Integer}}: the genotypes of the population at the end of the simulation.

locations::Vector{<:Any}: the locations of the population at the end of the simulation.

hybrid_zone_width::Real: the cline width associated with the male mating trait

population_overlap::Real: the proportion of the range occupied by males of both mating trait phenotypes

HZAM.DataAnalysis.PopulationTrackingData — Type

PopulationTrackingData

A PopulationTrackingData stores the size, hybridity, overlap, and male mating trait cline width of the population.

As the simulation runs a vector of PopulationTrackingData is used to keep track of the key population data each generation.

Fields

population_size::Real: the total number of individuals.

hybridity::Real: the average hybridity in the population on the male mating trait loci.

overlap::Real: the proportion of the range containing both male mating trait phenotypes.

width::Real: the cline width for the male mating trait.

Constructors

- PopulationTrackingData(
    genotypes::Vector{<:Matrix{<:Integer}},
    locations::Vector,
    male_mating_trait_loci::Union{UnitRange{<:Integer},Vector{<:Integer}}
)

HZAM.DataAnalysis.SimParams — Type

SimParams

A SimParams stores the parameters used in a simulation.

Fields

intrinsic_R::Real: the intrinsic growth rate.

ecolDiff::Real: the ecological difference between the two populations.

w_hyb::Real: the hybrid fitness.

S_AM::Real: the strength of assortative mating.

K_total::Integer: the carrying capacity of the environment.

max_generations::Integer: the number of generations the simulation ran for.

sigma_disp::Real: the standard deviation for dispersal distance.

total_loci::Integer: the number of loci in the genotypes.

female_mating_trait_loci: the loci controlling mating preference.

male_mating_trait_loci: the loci controlling mating cue.

competition_trait_loci: the loci controlling what resources the individual uses.

hybrid_survival_loci: the loci controlling hybrid fitness.

per_reject_cost::Real: the loss in fitness incurred by a female after rejecting one male.

HZAM.DataAnalysis.SpatialData — Type

SpatialData

A SpatialData stores the bimodality, gene flow, hybrid zone movement, cline width, and cline position data for the hybrid zone.

Fields

bimodality::Real: the proportion of phenotypically pure individuals within half a dispersal distance of the middle of the hybrid zone.

gene_flows::NamedTuple: the proportion of genetic material originating in the other species found in individuals above a phenotypic purity cutoff.

cline_widths::NamedTuple: the width of the cline for each trait.

cline_positions::NamedTuple: the position of the midpoint of the cline for each trait.

variance::Real: the average movement of the hybrid zone each generation.

Constructors

- SpatialData(
        locations::Vector,
        sigma_disp::Real,
        genotypes::Vector{<:Matrix{<:Integer}},
        loci::NamedTuple,
        last_generation::Bool
    )
- SpatialData(output_data::Vector{SpatialData})

Details on behaviour of different constructors

The first constructor computes the spatial data from the population's genotypes, standard deviation in dispersal distance, genotypes, and which loci control which traits.

The second constructor accepts a vector of SpatialData, computes the variance in the functional cline position, and the average over time of all remaining fields.

Functions

HZAM.DataAnalysis.average_data_per_phenotype — Method

average_data_per_phenotype(
    data::Vector{<:Integer},
    genotypes::Vector{<:Matrix{<:Integer}},
    loci::Union{UnitRange{<:Integer},Vector{<:Integer}}
)

Compute the average number of mates for males of each phenotype for a given loci range.

Arguments

data::Vector{<:Integer}: a list of the data to be averaged for each element in the genotypes vector

genotypes::Vector{<:Matrix{<:Integer}}: the genotypes for each individual

loci::Union{UnitRange{<:Integer},Vector{<:Integer}: the loci range governing the trait of interest.

HZAM.DataAnalysis.average_gene_data — Method

average_gene_data(gene_data::Vector{<:NamedTuple})

Compute the mean for each field of the given list of composite datatypes.

HZAM.DataAnalysis.average_width — Method

average_width(sigmoid_curves::Vector{<:Vector{<:Real}})

Compute the width of a cline given a series of sigmoid curves representing the cline along horizontal strips of the range.

HZAM.DataAnalysis.calc_all_cline_widths_and_positions — Method

calc_all_cline_widths_and_positions(
    genotypes::Vector{<:Matrix{<:Real}},
    locations::Vector,
    loci::NamedTuple
)

For each trait, compute the cline width and the location of the middle.

Arguments

genotypes::Vector{<:Matrix{<:Real}}: the genotypes of all individuals.

locations::Vector: the locations of all individuals.

loci::NamedTuple: the name and loci range of each trait of interest.

HZAM.DataAnalysis.calc_all_gene_flow — Method

calc_all_gene_flow(
    genotypes::Vector{<:Matrix{<:Real}},
    hybrid_indices_functional::Vector{<:Real},
    loci::NamedTuple
)

For each trait, compute the proportion of genetic material originating in the other species found in individuals above a phenotypic purity cutoff.

Arguments

genotypes::Vector{<:Matrix{<:Real}}: the genotypes of all individuals.

hybrid_indices_functional::Vector{<:Real}: the mean values of the genotypes over the functional loci only.

loci::NamedTuple: the name and loci range of each trait of interest.

HZAM.DataAnalysis.calc_all_trait_correlations — Method

calc_all_trait_correlations(genotypes::Vector{<:Matrix{<:Integer}}, loci::NamedTuple)

Compute the correlation between each trait.

Arguments

genotypes::Vector{<:Matrix{<:Integer}}: the genotype of every individual.

loci::NamedTuple: the name and loci range of each trait of interest.

HZAM.DataAnalysis.calc_average_linkage_diseq — Method

calc_average_linkage_diseq(genotypes::Vector{<:Matrix{<:Integer}}, loci::NamedTuple)

Compute the linkage disequilibrium between each loci and return a table of values of the average correlation between loci of each trait.

Arguments

genotypes::Vector{<:Matrix{<:Integer}}: the genotype of every individual.

loci::NamedTuple: the name and loci range of each trait of interest.

HZAM.DataAnalysis.calc_bimodality_in_range — Method

calc_bimodality_in_range(
    sigmoid_curve::Vector{<:Real},
    locations_x::Vector{<:Real},
    hybrid_indices::Vector{<:Real},
    sigma_disp::Real
)

Compute the proportion of phenotypically pure individuals within half a dispersal distance of the cline midpoint.

Arguments

sigmoid_curve::Vector{<:Real}: the sigmoid curve fitting the cline.

locations_x::Vector{<:Real}: the x values of the locations.

hybrid_indices::Vector{<:Real}: the mean values of the genotypes over the functional loci only.

sigma_disp::Real: the standard deviation in the dispersal distance distribution.

HZAM.DataAnalysis.calc_bimodality_overall — Method

Compute the bimodality across the entire range (see above for further description).

calc_bimodality_overall(
    sigmoid_curves::Vector{<:Vector{<:Real}},
    sorted_indices::Vector{<:Vector{<:Integer}},
    locations_x::Vector{<:Real},
    hybrid_indices::Vector{<:Real},
    sigma_disp::Real
)

Arguments

sigmoid_curves::Vector{<:Real}: the sigmoid curves for each horizontal strip of the range.

sorted_indices::Vector{<:Vector{<:Integer}}: the indices of the locations sorted into bins bins corresponding to non-overlapping ranges of the y values.

locations_x::Vector{<:Real}: the x values of the locations.

hybrid_indices::Vector{<:Real}: the mean values of the genotypes over the functional loci only.

sigma_disp::Real: the standard deviation in the dispersal distance distribution.

HZAM.DataAnalysis.calc_chi_squared — Method

calc_chi_squared(genotypes::Vector{<:Matrix{<:Integer}}, locus::Integer)

Compare the observed genotype frequencies at the given locus to those expected from Hardy-Weinberg analysis.

HZAM.DataAnalysis.calc_length — Method

calc_length(sigmoid_curves::Vector{<:Vector{<:Real}})

Compute the approximate length of the hybrid zone.

This is done by finding the midpoints of a series of sigmoid curves representing the cline along evenly spaced horizontal strips of the range. The cline length is defined as the length of the shortest path traversing the entire range passing through each midpoint.

HZAM.DataAnalysis.calc_linkage_diseq — Method

calc_linkage_diseq(
    genotypes::Vector{<:Matrix{<:Integer}},
    l1::Integer,
    l2::Integer
)

Compute the linkage disequilibrium (calculated using the Pearson coefficient) between two loci.

$r = \frac{\sum{(x_i -\bar{x})(y_i -\bar{y})}} {\sqrt{\sum{(x_i -\bar{x})^2}\sum{(y_i -\bar{y})^2}}}$

Arguments

genotypes::Vector{<:Matrix{<:Integer}}: the genotype of every individual.

l1::Integer: the first focal loci.

l2::Integer: the second focal loci.

HZAM.DataAnalysis.calc_overlap_in_range — Method

calc_overlap_in_range(
    locations_x::Vector{<:Real},
    hybrid_indices_functional::Vector{<:Real}
)

Compute the proportion of the range containing both species.

Subdivide the range into 50 boxes and sums the area covered by boxes containing at least 10% phenotypically pure individuals of both species.

HZAM.DataAnalysis.calc_overlap_overall — Method

function calc_overlap_overall(
    locations_x::Vector,
    hybrid_indices_functional::Vector{<:Real},
    sorted_indices::Vector{<:Vector{<:Integer}}
)

Compute the total overlap area between the two species.

Arguments

locations_x::Vector: the x values of all the locations.

hybrid_indices_functional::Vector{<:Real}: the mean values of the genotypes over the functional loci only.

sorted_indices::Vector{<:Vector{<:Integer}}: the indices of the locations sorted into

bins corresponding to non-overlapping ranges of the y values.

HZAM.DataAnalysis.calc_position — Method

calc_position(sigmoid_curves::Vector{<:Vector{T} where <:Real})

Calculate the x location at the middle of the cline based on a series of sigmoid curves fitting the cline at different ranges on the y axis.

HZAM.DataAnalysis.calc_sigmoid_curve — Method

calc_sigmoid_curve(locations_x::Vector{<:Real}, hybrid_indices::Vector{<:Real})

Compute a sigmoid curve that models hybrid index vs location on the x axis.

Return the output of the sigmoid curve function on 1000 evenly spaced locations.

HZAM.DataAnalysis.calc_sigmoid_curves — Method

calc_sigmoid_curves(locations::Vector, hybrid_indices::Vector{<:Real})

Divide the range into ten horizontal strips and compute a sigmoid curve fitting the hybrid indices along each strip.

HZAM.DataAnalysis.calc_trait_correlation — Method

calc_trait_correlation(
    genotypes::Vector{<:Matrix{<:Integer}},
    loci_range1::Union{UnitRange{<:Integer},Vector{<:Integer}},
    loci_range2::Union{UnitRange{<:Integer},Vector{<:Integer}}
)

Compute the correlation between two traits.

Arguments

genotypes::Vector{<:Matrix{<:Integer}}: the genotype of every individual.

loci_range1::Union{UnitRange{<:Integer},Vector{<:Integer}}: the loci range of the first trait.

loci_range2::Union{UnitRange{<:Integer},Vector{<:Integer}}: the loci range of the second trait.

HZAM.DataAnalysis.calc_traits_additive — Method

calc_traits_additive(
    genotypes::Vector{<:Matrix{<:Integer}},
    loci::Union{UnitRange{<:Integer},Vector{<:Integer}}
)::Vector{Float32}

Compute the mean values of the genotypes passed to it at the given loci. Used to determine trait values in an additive way.

HZAM.DataAnalysis.calc_variance — Method

calc_variance(positions::Vector{<:Real})

Calculate the average movement of the hybrid zone each generation.

HZAM.DataAnalysis.calc_width — Method

calc_width(sigmoid_curve::Vector{<:Real})

Compute the width of a cline given a sigmoid curve.

The cline width is defined as the distance between where the sigmoid curve passes through 0.1 and 0.9.

HZAM.DataAnalysis.count_phenotypes_at_loci — Method

function count_phenotypes_at_loci(
    genotypes::Vector{<:Matrix{<:Integer}},
    loci::Union{UnitRange{<:Integer},Vector{<:Integer}}
)

Return the number of individuals with each phenotype for the given loci range.

Arguments

genotypes::Vector{<:Matrix{<:Integer}}: the genotypes of every individual

loci::Union{UnitRange{<:Integer},Vector{<:Integer}}: the loci range over which the phenotypes are calculated.

HZAM.DataAnalysis.find_fixed_alleles — Method

find_fixed_alleles(genotypes::Vector{<:Matrix{<:Integer}})

Determine at which loci the population has lost its genetic diversity.

HZAM.DataAnalysis.sigmoid — Method

sigmoid(x::Vector{<:Real}, p::Vector{<:Real})

Compute the y value for a sigmoid given the x value, centre, and maximum slope.

Arguments

x::Real: the value along the x axis where the sigmoid is to be calculated.

p::Vector{<:Real}: the fit parameters where p[1] is the cline centre and p[2] is

the maximum slope.

HZAM.DataAnalysis.sort_locations — Method

sort_locations(A::AbstractArray{<:Real}, bin_size::Real)

Sort the given list of numbers into bins of a given size spanning 0 to 1.

HZAM.DataAnalysis.sort_locations_2D — Method

sort_locations_2D(locations::Vector{<:Any}, w::Integer)

Sort the indices of the locations into a w by w matrix with each cell in the matrix holding a vector of indices corresponding to the individuals in that (1/w) by (1/w) portion of the range.

HZAM.DataAnalysis.sort_y — Method

sort_y(y_locations::Vector{<:Real})

Sort the y coordinate indices into 10 vectors [0, 0.1), [0.1, 0.2), etc.

Example

jldoctest sort_y([0.01, 0.5, 0.24, 0.9]) 10-element Vector{<:Vector{Int64}}: [1] [] [3] [] [] [2] [] [] [] [4]