Population

HZAM.Population — Module

Types and functions used for managing population data (locations, genotypes, and growth rates).

Constants

HZAM.Population.NUM_ZONES — Constant

The width of the (square) grid of zones that divides the population into more manageable chunks. Default is 10x10.

Types

HZAM.Population.Location — Type

Location

A Location stores an x coordinate and a y coordinate.

Constructors

- Location(x::Float32, y::Float32)
- Location(starting_range::Vector{<:Location})
- Location(starting_location::Location, sigma_disp::Real)

Details on behaviour of different constructors

The first constructor creates a new location with the given x and y coordinates.

The second constructor generates a random location within the given range. Range is specified by a vector containing the coordinates of the bottom left and top right corners.

The third constructor generates an offspring's location based on the mother's location and the standard deviation of the dispersal distance.

HZAM.Population.PopulationData — Type

PopulationData

A PopulationData stores the genotypes, locations, and growth rates for all individuals in the simulation. The data is subdivided by 'zone' into a matrix for calculation efficiency.

Fields

population::Matrix{Zone}: all of the zones (containing the genotypes, locations, and resource use) in the simulation.

growth_rates_F::Matrix{Vector{Float64}}: the growth rates of every female for each zone.

Constructors

- PopulationData(
    K_total::Integer,
    ecolDiff::Real,
    total_loci::Integer,
    intrinsic_R::Real,
    sigma_comp::Real
)
- PopulationData(
    genotypes_daughters::Matrix{Vector{Matrix{Int8}}},
    genotypes_sons::Matrix{Vector{Matrix{Int8}}},
    locations_daughters::Matrix{Vector{Location}},
    locations_sons::Matrix{Vector{Location}},
    competition_trait_loci::Union{UnitRange{<:Integer},Vector{<:Integer}},
    K_total::Integer,
    sigma_comp::Real,
    intrinsic_R::Real,
    ecolDiff::Real
)

Details on behaviour of different constructors

The first constructor sets up the initial population for the simulation.

The second constructor creates a new PopulationData using the genotypes and locations of the offspring.

HZAM.Population.Zone — Type

Zone

A Zone stores the genotypes locations and resource use for the individuals contained within a subset of the range.

Fields

genotypes_F::Vector{Matrix{Int8}}: the female genotypes. rows are alleles (row 1 from mother, row 2 from father) and columns are loci.

genotypes_M::Vector{Matrix{Int8}}: the male genotypes.

locations_F::Vector{<:Location}: the female locations.

locations_M::Vector{<:Location}: the male locations.

ind_useResourceA_F::Vector{Float64}: each female's contribution to the use of resource A.

ind_useResourceA_M::Vector{Float64}: each male's contribution to the use of resource A.

ind_useResourceB_F::Vector{Float64}: each female's contribution to the use of resource B.

ind_useResourceB_M::Vector{Float64}: each male's contribution to the use of resource B.

Constructors

- Zone(
    starting_N::Integer,
    total_loci::Integer,
    location::Location,
    size::Float32,
    population::Real,
    ecolDiff::Real
)
- Zone(
    genotypes_F::Vector{Matrix{Int8}},
    genotypes_M::Vector{Matrix{Int8}},
    locations_F::Vector{<:Location},
    locations_M::Vector{<:Location},
    competition_trait_loci::Union{UnitRange{<:Integer},Vector{<:Integer}},
    ecolDiff
)

Details on behaviour of different constructors

The first constructor sets up the initial locations, genotypes, and resource use of every individual in the zone.

The second constructor creates a new Zone using the genotypes and locations of the offspring.

Functions

HZAM.Population.assign_zone — Method

assign_zone(location::Location)

Determine which zone a location falls in.

HZAM.Population.calc_growth_rates — Method

calc_growth_rates(
    population::Matrix{Zone},
    zone_index::CartesianIndex,
    K_total::Integer,
    sigma_comp::Real,
    intrinsic_R::Real
)

Compute the female growth rates.

Arguments

population::Matrix{Zone}: the matrix of zones storing all the locations, and genotypes in the simulation.

zone_index::CartesianIndex: the index of the focal zone.

K_total::Integer: the carrying capacity.

sigma_comp::Real: the standard deviation for the normal curve used in calculating local density.

intrinsic_R::Real: the intrinsic growth rate.

HZAM.Population.calc_ideal_densities — Method

calc_ideal_densities(
    K_total::Integer,
    sigma_comp::Real,
    locations_F::Vector{<:Location},
    max_dist::Real
)

Compute expected local densities, based on geographically even distribution of individuals at carrying capacity.

Densities are calculated using the following integral:

$K_{total}\int\limits_0^{2\pi}\int\limits_0^{0.03} \exp(-\frac{r^2}{2σ_{comp}^2})rdrdθ$

where the density function is equal to 0 outside of the range limits.

Arguments

K_total::Integer: the carrying capacity.

sigma_comp::Real: the standard deviation for the normal curve used in calculating local density.

locations_F::Vector{<:Location}: a list of the locations where the ideal density is to be computed.

max_dist::Real: the cutoff for the furthest away an individual can be and affect the density calculation. Should be 3x sigma_comp.

HZAM.Population.calc_ind_useResourceA — Method

calc_ind_useResourceA(competition_traits::Vector{Real}, ecolDiff::Real)

Compute the individual contributions to the use of resource A according to a linear gradient between the resource use of species 0 and species 1.

Arguments

competition_traits::Vector{Real}: the ecological competition trait values.

ecolDiff::Real: the ecological difference between the species.

HZAM.Population.calc_ind_useResourceB — Method

calc_ind_useResourceB(competition_traits::Vector{Real}, ecolDiff::Real)

Compute the individual contributions to the use of resource B according to a linear gradient between the resource use of species 0 and species 1.

Arguments

competition_traits::Vector{Real}: the ecological competition trait values.

ecolDiff::Real: the ecological difference between the species.

HZAM.Population.calc_real_densities — Method

calc_real_densities(
    neighbourhood::Matrix{Zone},
    locations_F::Vector{<:Location},
    max_dist::Real,
    sigma_comp::Real
)

Compute the population density at each female's location for both resources.

Arguments

neighbourhood::Matrix{Zone}: the zone indices for all the zones close enough to affect the density calculation.

locations_F::Vector{<:Location}: the locations of every female in the focal zone.

max_dist::Real: the distance cutoff for the density calculation.

sigma_comp::Real: the standard deviation used in the density calculation.

HZAM.Population.calc_squared_distances — Method

calc_squared_distances(location_list::Vector{<:Location}, focal_location::Location)

Compute the squared distances from a set of locations to a single location.

Example

julia> calc_squared_distances([Location(0.5f0, 0.5f0), Location(0.4f0, 0.4f0), 
Location(0f0,0f0)], Location(0.5f0, 0.5f0))
3-element Vector{Float32}:
 0.0
 0.019999998
 0.5

HZAM.Population.calc_survival_fitness_epistasis — Function

calc_survival_fitness_epistasis(
    genotype::Matrix{<:Integer}, 
    hybrid_survival_loci::Union{UnitRange{<:Integer},Vector{<:Integer}}, 
    w_hyb::Real, 
    beta::Real=1
)

Compute the survival fitness of an individual according to epistasis, with the beta parameter set to one as a default.

Arguments

genotype:Matrix{<:Integer}: the individual's genotype.

hybrid_survival_loci::Union{UnitRange{<:Integer},Vector{<:Integer}}: the loci responsible for heterozygote disadvantage in computing the survival fitness.

w_hyb::Real: the simulation's hybrid fitness value.

beta::Real

HZAM.Population.calc_survival_fitness_hetdisadvantage — Method

calc_survival_fitness_hetdisadvantage(genotype::Matrix{<:Integer}, w_hyb::Real)

Compute the survival fitness of an individual according to heterozygosity.

Arguments

genotype:Matrix{<:Integer}: the individual's genotype.

w_hyb::Real: the simulation's hybrid fitness value.

HZAM.Population.genotype_mean — Method

genotype_mean(genotype::Matrix{<:Integer}, loci::Union{UnitRange{<:Integer},Vector{<:Integer}})

Compute the mean value of the genotype across the list of loci given.

Example

julia> genotype_mean([0 1 0; 0 1 0], [1,3])
0.0

HZAM.Population.get_surrounding_zones — Method

get_surrounding_zones(zone_index::CartesianIndex)

Return a list of the indices of the neighbouring zones and the given zone index.

HZAM.Population.max_radius_squared — Method

max_radius_squared(x::Real, y::Real, t::Real, max_dist::Real)

Compute the distance from a point along an angle to the limit of the range (defined by x∈[0,1), y∈[0,1)).

The distance gets cut off at max_dist. Used in calculating the ideal densities.

Arguments

location::Location: the starting location.

t::Real: the angle.

max_dist: the maximum distance.